IMPROVEMENT-39: Move ssh modules to model. In progress
