+++ internal/model/sshconfig/config_test.go
@@ -1,4 +1,4 @@
-package ssh
+package sshconfig
 
 import (
 	"testing"
