+++ internal/model/utils.go
@@ -2,8 +2,9 @@ package model
 
 import (
 	"fmt"
-	"regexp"
 	"strings"
+
+	"github.com/grafviktor/goto/internal/utils"
 )
 
 // CommandLineOption - parent interface for command line option.
@@ -30,12 +31,6 @@ func constructKeyValueOption(optionFlag, optionValue string) string {
 	return ""
 }
 
-var twoOrMoreSpacesRegexp = regexp.MustCompile(`\s{2,}`)
-
-func removeDuplicateSpaces(arguments string) string {
-	return twoOrMoreSpacesRegexp.ReplaceAllLiteralString(arguments, " ")
-}
-
 func addOption(sb *strings.Builder, rawParameter CommandLineOption) {
 	var option string
 	switch p := rawParameter.(type) {
@@ -46,10 +41,10 @@ func addOption(sb *strings.Builder, rawParameter CommandLineOption) {
 	case OptionLoginName:
 		option = constructKeyValueOption("-l", p.Value)
 	case OptionReadConfig:
-		option = constructKeyValueOption("-G", removeDuplicateSpaces(p.Value))
+		option = constructKeyValueOption("-G", utils.RemoveDuplicateSpaces(p.Value))
 	case OptionAddress:
 		if p.Value != "" {
-			option = fmt.Sprintf(" %s", removeDuplicateSpaces(p.Value))
+			option = fmt.Sprintf(" %s", utils.RemoveDuplicateSpaces(p.Value))
 		}
 	default:
 		return
