+++ internal/state/state.go
@@ -8,7 +8,7 @@ import (
 
 	"gopkg.in/yaml.v2"
 
-	"github.com/grafviktor/goto/internal/utils/ssh"
+	"github.com/grafviktor/goto/internal/model/sshconfig"
 )
 
 type view int
@@ -39,11 +39,11 @@ type ApplicationState struct {
 	Selected         int `yaml:"selected"`
 	appStateFilePath string
 	logger           iLogger
-	CurrentView      view        `yaml:"-"`
-	Err              error       `yaml:"-"`
-	Width            int         `yaml:"-"`
-	Height           int         `yaml:"-"`
-	HostSSHConfig    *ssh.Config `yaml:"-"`
+	CurrentView      view              `yaml:"-"`
+	Err              error             `yaml:"-"`
+	Width            int               `yaml:"-"`
+	Height           int               `yaml:"-"`
+	HostSSHConfig    *sshconfig.Config `yaml:"-"`
 }
 
 // Get - reads application state from disk.
@@ -53,7 +53,7 @@ func Get(appHomePath string, lg iLogger) *ApplicationState {
 		appState = &ApplicationState{
 			appStateFilePath: path.Join(appHomePath, stateFile),
 			logger:           lg,
-			HostSSHConfig:    ssh.DefaultConfig(),
+			HostSSHConfig:    sshconfig.DefaultConfig(),
 		}
 
 		// If we cannot read previously created application state, that's fine - we can continue execution.
