+++ internal/ui/component/hostlist/hostlist.go
@@ -17,6 +17,7 @@ import (
 	"github.com/grafviktor/goto/internal/state"
 	"github.com/grafviktor/goto/internal/storage"
 	"github.com/grafviktor/goto/internal/ui/message"
+	"github.com/grafviktor/goto/internal/utils"
 )
 
 var (
@@ -340,7 +341,7 @@ func (m *listModel) listTitleUpdate() {
 	default:
 		// Replace Windows ssh prefix "cmd /c ssh" with "ssh"
 		newTitle = strings.Replace(item.Unwrap().CmdSSHConnect(), "cmd /c ", "", 1)
-		// newTitle = utils.RemoveDuplicateSpaces(newTitle)
+		newTitle = utils.RemoveDuplicateSpaces(newTitle)
 	}
 
 	if m.innerModel.Title != newTitle {
