+++ internal/ui/model.go
@@ -11,13 +11,14 @@ import (
 	"github.com/charmbracelet/bubbles/viewport"
 	tea "github.com/charmbracelet/bubbletea"
 
+	"github.com/grafviktor/goto/internal/model"
+	"github.com/grafviktor/goto/internal/model/sshconfig"
 	"github.com/grafviktor/goto/internal/state"
 	"github.com/grafviktor/goto/internal/storage"
 	"github.com/grafviktor/goto/internal/ui/component/hostedit"
 	"github.com/grafviktor/goto/internal/ui/component/hostlist"
 	"github.com/grafviktor/goto/internal/ui/message"
 	"github.com/grafviktor/goto/internal/utils"
-	"github.com/grafviktor/goto/internal/utils/ssh"
 )
 
 type iLogger interface {
@@ -101,7 +102,7 @@ func (m *mainModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		return m, m.dispatchProcessSSHLoadConfig(msg)
 	case message.RunProcessSuccess:
 		if msg.ProcessName == "ssh_load_config" {
-			parsedSSHConfig := ssh.ParseConfig(*msg.Output)
+			parsedSSHConfig := sshconfig.ParseConfig(*msg.Output)
 			m.appState.HostSSHConfig = parsedSSHConfig
 			m.logger.Debug("[UI] Update app state with host SSH config: %+v", *parsedSSHConfig)
 			cmds = append(cmds, message.TeaCmd(message.HostSSHConfigLoaded{}))
@@ -236,7 +237,7 @@ func (m *mainModel) dispatchProcess(name string, process *exec.Cmd, inBackground
 
 func (m *mainModel) dispatchProcessSSHConnect(msg message.RunProcessConnectSSH) tea.Cmd {
 	m.logger.Debug("[EXEC] Build ssh connect command for hostname: %v, title: %v", msg.Host.Address, msg.Host.Title)
-	process := utils.BuildConnectSSH(&msg.Host)
+	process := model.BuildConnectSSH(&msg.Host)
 	m.logger.Info("[EXEC] Run process: %s", process.String())
 
 	return m.dispatchProcess("ssh_connect_host", process, false, false)
@@ -244,7 +245,7 @@ func (m *mainModel) dispatchProcessSSHConnect(msg message.RunProcessConnectSSH) 
 
 func (m *mainModel) dispatchProcessSSHLoadConfig(msg message.RunProcessLoadSSHConfig) tea.Cmd {
 	m.logger.Debug("[EXEC] Read ssh configuration for host: %v", msg.Host)
-	process := utils.BuildLoadSSHConfig(&msg.Host)
+	process := model.BuildLoadSSHConfig(&msg.Host)
 	m.logger.Info("[EXEC] Run process: %s", process.String())
 
 	// Should run in non-blocking fashion for ssh load config
